build image & move tx metrics into chainClient
